Move state management for sidebar visibility inside of sidebar
